+++ src/widgets/settingspages/GeneralPage.cpp
@@ -1285,6 +1285,13 @@ void GeneralPage::initLayout(GeneralPageView &layout)
             R"(Pronouns are retrieved from <a href="https://pr.alejo.io">pr.alejo.io</a> when a user card is opened.)")
         ->addTo(layout);
 
+    SettingWidget::checkbox("Show stream title in live message",
+                            s.showTitleInLiveMessage)
+        ->setTooltip("The title in the message will be the title the streamer "
+                     "set when they went live, and will not update as the "
+                     "streamer updates their title.")
+        ->addTo(layout);
+
     SettingWidget::checkbox("Bold @usernames", s.boldUsernames)
         ->setTooltip("Bold @mentions to make them more noticeable.")
         ->addTo(layout);